app/requirements.txt
argcomplete==1.12.3
backcall==0.2.0
bleach==4.1.0
Brotli==1.0.9
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
cryptography==35.0.0
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.4.3
decorator==5.1.0
docutils==0.17.1
entrypoints==0.3
Flask==2.0.1
Flask-Compress==1.10.1
google-api-core==2.0.1
google-auth==2.2.1
google-cloud-automl==2.4.2
googleapis-common-protos==1.53.0
grpcio==1.41.0
idna==3.2
importlib-metadata==4.8.1
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.1
jupyter-client==7.0.5
jupyter-core==4.8.1
keyring==23.2.1
keyrings.google-artifactregistry-auth==0.0.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.1
numpy==1.21.2
packaging==21.0
pandas==1.3.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pkginfo==1.7.1
plotly==5.3.1
pluggy==1.0.0
prompt-toolkit==3.0.20
proto-plus==1.19.2
protobuf==3.18.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
readme-renderer==30.0
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rsa==4.7.2
SecretStorage==3.3.1
six==1.16.0
tenacity==8.0.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
twine==3.4.2
typing-extensions==3.10.0.2
urllib3==1.26.7
waitress==2.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
zipp==3.6.0
